+++ CsdnReader2/CsdnReader2/Controllers/CategoryArticlesViewController.m
@@ -16,7 +16,7 @@
 {
     FLTableView *_tableView;
     NSArray *articleLists;
-    NSInteger page;
+    int page;
 }
 
 @end
